Rename pr95 test directory to pr202
