+++ internal/testdata/snapshots/input/pr202/go.mod
@@ -1,0 +1,3 @@
+module sg/pr202
+
+go 1.19
