+++ internal/testdata/snapshots/input/pr202/local_anon.go
@@ -1,4 +1,4 @@
-package pr95
+package pr202
 
 // Anonymous structs in local/function scope.
