+++ internal/testdata/snapshots/input/pr202/multiname_fields.go
@@ -1,4 +1,4 @@
-package pr95
+package pr202
 
 // Multi-name field declarations: a, b share a type and must be siblings.
