--- internal/testdata/snapshots/input/pr95/go.mod
@@ -1,3 +1,0 @@
-module sg/pr95
-
-go 1.19
