+++ internal/testdata/snapshots/output/pr202/multiname_fields.go
@@ -1,10 +1,10 @@
-  package pr95
-//        ^^^^ definition 0.1.test `sg/pr95`/
+  package pr202
+//        ^^^^^ definition 0.1.test `sg/pr202`/
   
   // Multi-name field declarations: a, b share a type and must be siblings.
   
   type MultiNameStruct struct {
-//     ^^^^^^^^^^^^^^^ definition 0.1.test `sg/pr95`/MultiNameStruct#
+//     ^^^^^^^^^^^^^^^ definition 0.1.test `sg/pr202`/MultiNameStruct#
 //     documentation
 //     > ```go
 //     > type MultiNameStruct struct
@@ -23,24 +23,24 @@
 //     > }
 //     > ```
    a, b struct {
-// ^ definition 0.1.test `sg/pr95`/MultiNameStruct#a.
+// ^ definition 0.1.test `sg/pr202`/MultiNameStruct#a.
 // documentation
 // > ```go
 // > struct field a struct{x int; y string}
 // > ```
-//    ^ definition 0.1.test `sg/pr95`/MultiNameStruct#b.
+//    ^ definition 0.1.test `sg/pr202`/MultiNameStruct#b.
 //    documentation
 //    > ```go
 //    > struct field b struct{x int; y string}
 //    > ```
     x int
-//  ^ definition 0.1.test `sg/pr95`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
+//  ^ definition 0.1.test `sg/pr202`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
 //  documentation
 //  > ```go
 //  > struct field x int
 //  > ```
     y string
-//  ^ definition 0.1.test `sg/pr95`/MultiNameStruct#$anon_c0a8952b3a214f68#y.
+//  ^ definition 0.1.test `sg/pr202`/MultiNameStruct#$anon_c0a8952b3a214f68#y.
 //  documentation
 //  > ```go
 //  > struct field y string
@@ -48,29 +48,29 @@
    }
   }
   
-//⌄ enclosing_range_start 0.1.test `sg/pr95`/useMultiNameFields().
+//⌄ enclosing_range_start 0.1.test `sg/pr202`/useMultiNameFields().
   func useMultiNameFields() {
-//     ^^^^^^^^^^^^^^^^^^ definition 0.1.test `sg/pr95`/useMultiNameFields().
+//     ^^^^^^^^^^^^^^^^^^ definition 0.1.test `sg/pr202`/useMultiNameFields().
 //     documentation
 //     > ```go
 //     > func useMultiNameFields()
 //     > ```
    var m MultiNameStruct
 //     ^ definition local 0
-//       ^^^^^^^^^^^^^^^ reference 0.1.test `sg/pr95`/MultiNameStruct#
+//       ^^^^^^^^^^^^^^^ reference 0.1.test `sg/pr202`/MultiNameStruct#
    m.a.x = 1
 // ^ reference local 0
-//   ^ reference 0.1.test `sg/pr95`/MultiNameStruct#a.
-//     ^ reference 0.1.test `sg/pr95`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
+//   ^ reference 0.1.test `sg/pr202`/MultiNameStruct#a.
+//     ^ reference 0.1.test `sg/pr202`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
    m.b.x = 2
 // ^ reference local 0
-//   ^ reference 0.1.test `sg/pr95`/MultiNameStruct#b.
-//     ^ reference 0.1.test `sg/pr95`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
+//   ^ reference 0.1.test `sg/pr202`/MultiNameStruct#b.
+//     ^ reference 0.1.test `sg/pr202`/MultiNameStruct#$anon_c0a8952b3a214f68#x.
    m.a = m.b
 // ^ reference local 0
-//   ^ reference 0.1.test `sg/pr95`/MultiNameStruct#a.
+//   ^ reference 0.1.test `sg/pr202`/MultiNameStruct#a.
 //       ^ reference local 0
-//         ^ reference 0.1.test `sg/pr95`/MultiNameStruct#b.
+//         ^ reference 0.1.test `sg/pr202`/MultiNameStruct#b.
   }
-//⌃ enclosing_range_end 0.1.test `sg/pr95`/useMultiNameFields().
+//⌃ enclosing_range_end 0.1.test `sg/pr202`/useMultiNameFields().
